Be smarter about checking the collectd module.
Collectd's python api works by exposing a module called "collectd",
which unfortunately conflicts with a completely separate module
available on pypi called "collectd" as well.

In order to determine that the embeded collectd module is present we
just check if register_config() is present and watch for the
AttributeError if it isn't.

Fixes #1